+++ examples/Book/JsonApi/Document/BookDocument.php
@@ -5,7 +5,7 @@ namespace WoohooLabs\Yin\Examples\Book\JsonApi\Document;
 
 use WoohooLabs\Yin\Examples\Book\JsonApi\Resource\BookResourceTransformer;
 use WoohooLabs\Yin\JsonApi\Document\AbstractSingleResourceDocument;
-use WoohooLabs\Yin\JsonApi\Schema\JsonApi;
+use WoohooLabs\Yin\JsonApi\Schema\JsonApiObject;
 use WoohooLabs\Yin\JsonApi\Schema\Link;
 use WoohooLabs\Yin\JsonApi\Schema\Links;
 
@@ -22,11 +22,11 @@ class BookDocument extends AbstractSingleResourceDocument
      * The method returns a new JsonApi schema object if this member should be present or null
      * if it should be omitted from the response.
      *
-     * @return JsonApi|null
+     * @return JsonApiObject|null
      */
     public function getJsonApi()
     {
-        return new JsonApi("1.0");
+        return new JsonApiObject("1.0");
     }
 
     /**